+++ packages/mutator-specification/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.3](https://github.com/stryker-mutator/stryker/compare/v1.0.2...v1.0.3) (2019-02-26)
+
+**Note:** Version bump only for package @stryker-mutator/mutator-specification
+
+
+
+
+
 ## [1.0.0](https://github.com/stryker-mutator/stryker/compare/stryker-mutator-specification@0.7.1...@stryker-mutator/mutator-specification@1.0.0) (2019-02-13)