+++ packages/util/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.3](https://github.com/stryker-mutator/stryker/compare/v1.0.2...v1.0.3) (2019-02-26)
+
+**Note:** Version bump only for package @stryker-mutator/util
+
+
+
+
+
 # [0.1.0](https://github.com/stryker-mutator/stryker/compare/@stryker-mutator/util@0.0.3...@stryker-mutator/util@0.1.0) (2019-02-08)